+++ calign.h
@@ -2,10 +2,10 @@
 #include <stdbool.h>
 
 typedef struct align_t {
-    int s;
+    short s;
     size_t len;
     short* a1;
     short* a2;
 } align_t;
 
-align_t align(size_t len_a, const short* a, size_t len_b, const short* b, int d, size_t len_S, const int* S, bool local);
+align_t align(size_t len_a, const short* a, size_t len_b, const short* b, short d, size_t len_S, const short* S, bool local);
